Extract filters dialog out to seperate component & add actions
